Move stripes-core tests to miragejs
