+++ test/bigtest/helpers/setup-core-application.js
@@ -1,0 +1,6 @@
+import setupApplication from './setup-application';
+
+export default function setupCoreApplication(options = {}) {
+  options.mirageOptions = { serverType: 'miragejs' };
+  setupApplication(options);
+}
